Use fullmatch for regex + more tests
